New main with my file change
Just doing it here again. I made a spinner that uses the number of clock cycles the spin button was held down for to randomly select a final position on the wheel. It simulates the wheel slowing down before it stopped. I assumed that the lights were on bits 0 to 7 of PORTA, but because my calculations were done in a register this can be easily modified.